[ST-Engine] Add PipelineBaseScheduler For Engine (#2396)
* Add SeaTunnel Engine ConfigProvider and seatunnel-seatunnel-starter

* add source file to licenserc.yaml

* fix checkstyle

* fix error

* tmp

* tmp

* Update PhysicalPlan to support scheduler by pipeline

* remove init in run

* tmp

* complete scheduler

* optimize scheduler

* fix checkstyle

* fix checkstyle

* Update seatunnel-core/seatunnel-seatunnel-starter/src/main/java/org/apache/seatunnel/core/starter/seatunnel/SeaTunnelStarter.java


* Update SeaTunnelStarter.java

+++ seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/dag/actions/AbstractAction.java
@@ -27,20 +27,20 @@ public abstract class AbstractAction implements Action {
     private String name;
     private List<Action> upstreams = new ArrayList<>();
     // This is used to assign a unique ID to every Action
-    private int id;
+    private long id;
 
     private int parallelism = 1;
 
     private List<URL> jarUrls;
 
-    protected AbstractAction(int id, @NonNull String name, @NonNull List<Action> upstreams, @NonNull List<URL> jarUrls) {
+    protected AbstractAction(long id, @NonNull String name, @NonNull List<Action> upstreams, @NonNull List<URL> jarUrls) {
         this.id = id;
         this.name = name;
         this.upstreams = upstreams;
         this.jarUrls = jarUrls;
     }
 
-    protected AbstractAction(int id, @NonNull String name, @NonNull List<URL> jarUrls) {
+    protected AbstractAction(long id, @NonNull String name, @NonNull List<URL> jarUrls) {
         this.id = id;
         this.name = name;
         this.jarUrls = jarUrls;
@@ -79,7 +79,7 @@ public abstract class AbstractAction implements Action {
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return id;
     }
